+++ design-system/page-layout/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page-layout
 
+## 0.9.4
+
+### Patch Changes
+
+- [`58413f42ab`](https://bitbucket.org/atlassian/atlassian-frontend/commits/58413f42ab) - Left side bar button disbaled
+
 ## 0.9.3
 
 ### Patch Changes