+++ editor/embedded-document/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/embedded-document
 
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.7.4
 
 ### Patch Changes